move models into models module
