+++ helpers/models/sd3/pipeline.py
