+++ helpers/models/sdxl/pipeline.py
